Fixed a class cast exception in loadModel()
git-svn-id: https://cdk.svn.sourceforge.net/svnroot/cdk/trunk/cdk@4611 eb4e18e3-b210-0410-a6ab-dec725e4b171

+++ src/org/openscience/cdk/qsar/model/R/CNNRegressionModel.java
@@ -470,8 +470,8 @@ public class CNNRegressionModel extends RModel {
         if (model.getClass().getName().equals("org.openscience.cdk.qsar.model.R.CNNRegressionModelFit")) {
             this.modelfit = (CNNRegressionModelFit)model;
             this.setModelName(modelName);
-            Double tmp = (Double)revaluator.eval(modelName+"$n[1]");
-            nvar = (int)tmp.doubleValue();
+            Integer tmp = (Integer)revaluator.eval(modelName+"$n[1]");
+            nvar = tmp.intValue();
         } else throw new QSARModelException("The loaded model was not a CNNRegressionModel");
     }
     /**
